Bug 1766646 - (fix-9a99905301f) fix a casting issue

+++ video/video_stream_buffer_controller.cc
@@ -364,7 +364,8 @@ void VideoStreamBufferController::UpdateTimingFrameInfo() {
 
 bool VideoStreamBufferController::IsTooManyFramesQueued() const
     RTC_RUN_ON(&worker_sequence_checker_) {
-  return buffer_->CurrentSize() > zero_playout_delay_max_decode_queue_size_;
+  return buffer_->CurrentSize() >
+         static_cast<unsigned int>(zero_playout_delay_max_decode_queue_size_);
 }
 
 void VideoStreamBufferController::ForceKeyFrameReleaseImmediately()
